A ton of updates to the renderer.
Now we can correctly set up and clear our rendertargets.

Reviewed-on: https://wnproject.org/gerrit/300
Try: AWoloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>

+++ Overlays/Vulkan/Runtime/WNGraphics/inc/Internal/Vulkan/WNDevice.h
@@ -121,7 +121,7 @@ public:
   };
 
   math::vec2f get_2d_transform_scale() WN_GRAPHICS_OVERRIDE_FINAL {
-    return math::vec2f({ 1.0f, -1.0f });
+    return math::vec2f({1.0f, -1.0f});
   }
 
 protected:
@@ -152,8 +152,8 @@ protected:
       const clear_value& _optimized_clear,
       image* _image) WN_GRAPHICS_OVERRIDE_FINAL;
   void destroy_image(image* _image) WN_GRAPHICS_OVERRIDE_FINAL;
-  void bind_image_memory(
-      image* _image, arena* _arena, size_t _offset) WN_GRAPHICS_OVERRIDE_FINAL;
+  void bind_image_memory(image* _image, arena* _arena,
+      uint64_t _offset) WN_GRAPHICS_OVERRIDE_FINAL;
   image_memory_requirements get_image_memory_requirements(
       const image* _image) WN_GRAPHICS_OVERRIDE_FINAL;
 
@@ -234,7 +234,8 @@ protected:
   void destroy_sampler(sampler* _sampler) WN_GRAPHICS_OVERRIDE_FINAL;
 
   // arena methods
-  bool initialize_arena(arena* _arena, const size_t _index, const size_t _size,
+  bool initialize_arena(arena* _arena, const size_t _index,
+      const uint64_t _size,
       const bool _multisampled) WN_GRAPHICS_OVERRIDE_FINAL;
   void destroy_arena(arena* _arena) WN_GRAPHICS_OVERRIDE_FINAL;
